[Fix] #560 - Home, soptlog coor에서 protocol과 enum 분리
