+++ SOPT-iOS/Projects/Features/SoptlogFeature/Interface/Sources/SoptlogCoordinatorDestination.swift
@@ -1,0 +1,15 @@
+//
+//  SoptlogCoordinatorDestination.swift
+//  SoptlogFeatureInterface
+//
+//  Created by Jae Hyun Lee on 5/6/25.
+//  Copyright © 2025 SOPT-iOS. All rights reserved.
+//
+
+import Foundation
+
+public enum SoptlogCoordinatorDestination {
+    case dailySoptune
+    case signIn
+    case webLink(url: String)
+}
