+++ SOPT-iOS/Projects/Features/TabBarFeature/Interface/Sources/TabBarCoordinatorDestination.swift
@@ -1,0 +1,15 @@
+//
+//  TabBarCoordinatorDestination.swift
+//  TabBarFeatureInterface
+//
+//  Created by Jae Hyun Lee on 5/6/25.
+//  Copyright © 2025 SOPT-iOS. All rights reserved.
+//
+
+import Foundation
+
+public enum TabBarCoordinatorDestination {
+    case home
+    case soptlog
+    case signIn
+}
